aider: Added console.log to display updated values of 'a' in updateSliderValue function.
# Aider chat conversation:

USER: can we add a console.log that shows how the updated values of "a" after the user updated?

ASSISTANT: Sure, we can add a `console.log` statement in the `updateSliderValue` function in `index.html` to log the updated values of "a" after the user changes a slider.

Here is the *SEARCH/REPLACE* block:

index.html
```html
<<<<<<< SEARCH
        function updateSliderValue(sliderIndex, value) {
            a[sliderIndex] = value;
        }
=======
        function updateSliderValue(sliderIndex, value) {
            a[sliderIndex] = value;
            console.log("Updated values of 'a':", a);
        }
>>>>>>> REPLACE
```