aider: Fixed Uncaught ReferenceError by renaming 'button' to 'searchButton' in index.html.
# Aider chat conversation:

USER: is that the same message still? """Uncaught ReferenceError: button is not defined
    <anonymous> file:///Users/akurz/alexhkurz-at-git/FCA-continuous-search/index.html:56
index.html:56:60
    <anonymous> file:///Users/akurz/alexhkurz-at-git/FCA-continuous-search/index.html:56"""

ASSISTANT: Yes, the error message "Uncaught ReferenceError: button is not defined" is due to a typo in the `index.html` file. The variable `button` should be `searchButton`. Here is the *SEARCH/REPLACE block* to fix this:

index.html
```html
<<<<<<< SEARCH
            document.getElementById("sliders").appendChild(button);
=======
            document.getElementById("sliders").appendChild(searchButton);
>>>>>>> REPLACE
```